fixed last bug of the potetiometer, and in the CUS the send task

+++ CUS/src/http/MeasurementsHandler.java
@@ -36,7 +36,7 @@ public class MeasurementsHandler implements HttpHandler {
                 n = Integer.parseInt(q.substring(5));
             }
 
-            List<Integer> values =
+            List<Float> values =
                     controller.getMeasurements(n);
 
             //build the json with all the measurments
